Set watches when initializing, and remove needless loop
The initializing isn't usually necessary, but I sourced
`nodeinspect.vim` when tracking down other issues, which broke the
watches because then `s:session["watches"]` didn't exist any more.

If we already set s:session["watches"] = {} on startup, there's no
reason to not set it in the init that I can see. I left the set in
loadSessionFile() in as well, in case there's some edge case where we
are loading a session file and don't want to merge with some existing
session watches or something.

Also, we were unnecessarily looping through the keys and calling
AddBulk() with the same argument multiple times. It was idempotent so it
was fine, but it's also not useful, and I imagine left over from some
previous restructuring.